Raise error before jobs getting stuck in  CREATING STACK cfn-airflow-flower
If your job is stuck here for more then 10 minutes, you probably have an issue with your deployment. The main causes for this is ECS tasks not being able to find the image definition on ECR, possibly caused by an autentication error between Docker and ECR.

Now if there was an issue while pushing the image, we will raise an error instead of continuing deployment.